💩 src: Code refactoring then minor cleanup
* Rewrote `calendar.ts` with just ` for-loop to do the job. Feel free to
  submit optimization PR to improve the perf.
* Minor code re-ordering for maintainability and readability.
* Removed other test scripts which are no longer in use.
