+++ index.html
@@ -15,7 +15,7 @@
   <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Google+Sans:400,500,700">
   <script type="module">
     import '/dist/app-datepicker.js';
-    import '/dist/test-rerender.js';
+    // import '/dist/test-rerender.js';
     // import '/dist/calendar-benchmark.js';
   </script>
 </head>
